Refactor: background script 모듈화
