+++ extension/manifest.json
@@ -6,7 +6,7 @@
   "permissions": ["storage", "activeTab", "scripting"],
   "host_permissions": ["https://solved.ac/problems*"],
   "background": {
-    "service_worker": "src/background/script.ts",
+    "service_worker": "src/services/chrome/background-script.ts",
     "type": "module"
   },
   "content_scripts": [
