--- extension/src/background/index.ts
@@ -1,1 +1,0 @@
-export { type BackgroundMessage, type BackgroundResponse } from './script'
