+++ extension/src/services/chrome/background/index.ts
@@ -1,0 +1,1 @@
+export { BackgroundService } from './background-service'
