+++ extension/src/types/index.ts
@@ -1,2 +1,3 @@
 export { type SolvedAcUser, type SolvedAcProblem } from './solved-ac'
 export { type User } from './common'
+export { type BackgroundMessage, type BackgroundResponse } from './background'
